+++ src/domains/record/mocks/create-record-detail.ts
