+++ src/domains/record/mocks/delete-record-detail.ts
