+++ src/domains/record/mocks/patch-record-detail.ts
