+++ src/domains/record/mocks/post-summary-record-detail.ts
